Check if customer exists if they don't then redirect them to pricing page
